refactor(shaders): extract shading-change diff metric into a shared header
Move the per-pixel shading-change difference metric (the sample sets, sorting
network, ComputeDiff, and SpdLoadSourceImage) out of
ffx_fsr3upscaler_shading_change_pyramid.h into a new
ffx_fsr3upscaler_shading_change_diff.h, included by the pyramid header. This
lets the upcoming Core write-only mip0 pass share the exact same metric.
Preprocessed output is unchanged, so the compiled SPIR-V is identical.
